requirements.txt
pymongo==2.4.2
bottle==0.11.6
argparse==1.2.1
GitPython==0.1.7
PyGithub==1.25.2
requests>=2.4.3
PyYAML>=3.11

